modularized the Overview screen in 5 composables separate in : topbar,bottombar,content,trip and one regrouping everything with also the overviewViewModel
