+++ app/src/main/java/com/github/se/wanderpals/ui/screens/overview/Overview.kt
@@ -1,0 +1,99 @@
+package com.github.se.wanderpals.ui.screens.overview
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+
+import androidx.compose.foundation.layout.Column
+
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Clear
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Share
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.DockedSearchBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.github.se.wanderpals.model.data.Trip
+import com.github.se.wanderpals.model.viewmodel.OverviewViewModel
+import com.github.se.wanderpals.ui.navigation.NavigationActions
+import com.github.se.wanderpals.ui.navigation.Route
+import java.time.format.DateTimeFormatter
+
+
+
+@Composable
+fun Overview(overviewViewModel: OverviewViewModel, navigationActions: NavigationActions) {
+
+    val tripsList by overviewViewModel.state.collectAsState()
+    val isLoading by overviewViewModel.isLoading.collectAsState()
+
+    val EMPTY_SEARCH = ""
+    var searchText by remember { mutableStateOf(EMPTY_SEARCH) }
+
+
+    if (isLoading) {
+        Box(modifier = Modifier.fillMaxSize()) {
+            CircularProgressIndicator(
+                modifier = Modifier
+                    .size(50.dp)
+                    .align(Alignment.Center)
+            )
+        }
+    } else {
+        Scaffold(
+            topBar = {
+                OverviewTopBar(
+                    searchText = searchText,
+                    onSearchTextChanged = { newSearchText -> searchText = newSearchText }
+                )
+            },
+            bottomBar = {
+                OverviewBottomBar(
+                    onCreateTripClick = { navigationActions.navigateTo(Route.CREATE_TRIP) },
+                    onLinkClick = {/*TODO do this when starting the link invitation*/ })
+            }
+        ) {
+            innerPadding ->
+            OverviewContent(
+                innerPadding = innerPadding,
+                navigationActions = navigationActions,
+                tripsList = tripsList,
+                searchText = searchText)
+        }
+
+    }
+
+}
